refactor: Model의 더미데이터 및 비즈니스 로직을 뷰 모델로 이동
