+++ SwiftUICoinExchange/View/Lectures/231127/HorizontalViewModel.swift
@@ -11,10 +11,32 @@ import Foundation
 final class HorizontalViewModel: ObservableObject {
    
     @Published var time = 0.0 // View가 아니므로 @State가 아닌 @Published
+    @Published var dummy: [Currency] = []
     
     func timer() {
         Timer.scheduledTimer(withTimeInterval: 0.5, repeats: true) { timer in
             self.time += 0.5
+            self.fetchDummyData() // 0.5초마다 데이터 변화
+        }
+    }
+    
+    func fetchDummyData() {
+        
+        dummy = [
+            Currency(name: "ANDY"),
+            Currency(name: "NOAH"),
+            Currency(name: "SSABOO"),
+            Currency(name: "JOEY"),
+            Currency(name: "CALLIE"),
+            Currency(name: "YERI")
+        ]
+        
+        func showlargest() -> Int {
+            let data = dummy.sorted { $0.priceInUSD > $1.priceInUSD
+            }
+            return data[0].priceInUSD
         }
     }
 }
+
+
